requirements.txt
# Core
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# ML & Statistics
numpy>=1.26.0,<2.0.0
scipy>=1.12.0,<2.0.0
scikit-learn>=1.4.0,<2.0.0

# Monitoring
evidently>=0.4.0

# Storage
psycopg2-binary>=2.9.9,<3.0.0
redis>=5.0.0,<6.0.0

# Metrics
prometheus-client>=0.20.0,<1.0.0

# HTTP
httpx>=0.27.0,<1.0.0

# Configuration
pyyaml>=6.0.1,<7.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
pytest-asyncio>=0.23.0,<1.0.0

# Linting
ruff>=0.3.0
